ml-service/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
timm
pillow
numpy